+++ src/main/java/com/root/whattodotoday/member/domain/Member.java
@@ -1,6 +1,5 @@
 package com.root.whattodotoday.member.domain;
 
-import com.root.whattodotoday.todo.domain.TodoStatus;
 import lombok.Getter;
 
 import javax.persistence.*;
@@ -12,7 +11,7 @@ public class Member {
     @Id
     @GeneratedValue
     @Column(name = "member_no")
-    private int memberNo;
+    private Long memberNo;
 
     private String id;
     private String pw;
